requirements.txt
requests
pywin32
pycrypto